requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic-ai-slim[openai]
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
faster-whisper
python-multipart
certifi
